+++ spec/app/models/band.rb
@@ -7,6 +7,7 @@ class Band
   field :member_count, type: Integer
   field :members, type: Array
   field :likes, type: Integer
+  field :views, type: Integer
 
   embeds_many :records, cascade_callbacks: true
   embeds_many :notes, as: :noteable, cascade_callbacks: true, validate: false